requirements.txt
Flask
pandas
openpyxl
python-dotenv
requests
plotly